Avoid running migrations on select-only database
There is an inconsistency in Sequel's migrator classes, where the
Integer migrator always attempts a CREATE IF NOT EXISTS on the schema
information table (whereas the Timestamp migrator checks with a select).
Since we are using the Integer migrator, we have to check a flag
ourselves to support using a select-only database.