AEIM-873 - Resolve client attributes when direct or proxied
This introduces two new classes to resolve the request attributes of the
client. They are called ClientResolver and ProxiedResolver for now,
though those may not be the best names long-term.

These are each very simple, extracting a username and client_ip from the
request. There is a basic mechanism on the Keycard module to set which
to use, which the application should do based on configuration. We may
choose to do configuration reading within Keycard or make this into a
flag where the choice of class is internal to Keycard based on a fixed
set of options.