AEIM-873 - Use wrapped requests instead of resolvers
This change clarifies responsibilities and addresses a small design
consideration of how the client IP is determined for use by the
InstitutionFinder. Rather than the previous design, which introduced a
concept of a "resolver" for the user/client attributes, there are now
different types of wrappers for the base (Rails or Rack) request. The
interface for a request wrapper is that it has two methods: `username`
and `client_ip`. Now, any additional collaborators that need to extract
attributes from the request will have these two properties available.

This change also accounts for multiple remote addresses and adds some
configuration information to the README.