WIP
shib attribute mapping - do not use (in proxied request spec) - this
should really go in a separate finder that Noah will make a subtask
to add

supplemental vs. identity attributes - started this but the way the
specs & requestattributes work with the original Rack request and the
proxied vs. direct request vs. the actual attribute extraction and
mocking all of that in the tests is a little bit confusing