AEIM-1315 - Move user_attributes back to the request
Initially, the goal was to turn the finders into a list, such that we
would map and merge over them. A finder for the user attributes would be
the first new one. I started that work, but realized that there was
already a place for exactly that work: the Direct and Proxied requests
themselves. Which attributes are extracted is specific to the access
type, so this change adds an attributes method there.

There are a couple of things that this uncovers:

 * We need to decide between string and symbol keys for the attributes,
   since there is a bit of a mix going on now.
 * The name `attributes` is a little strange for the method called on an
   instance named `request`, given that we are doing so in a class
   called RequestAttributes. Maybe it's really `user_attributes`, or
   maybe this needs a small step back. I do like that there is a single
   entry point that dispatches to the right request type based on
   config, but it's possible that there is a base class/factory hiding.