AEIM-1320 - Extract Shibboleth attributes
This change takes a little bit of liberty with how the attributes are
aliased for use downstream of Keycard. I've tried to come up with what
feel reasonably accurate, usable, and resilient keys for the base
attribute set. The more official attribute names are exposed under a
special "verbatim" method here.