AIEM-1320 - Make Shibboleth mapping more direct
Instead of doing aliasing/indirection on the attributes, we decided to
map them directly for now. For example, instead of affiliation, we are
using eduPersonScopedAffiliation. This is to avoid premature or
inaccurate abstractions before we know that we need to cover multiple
SSO systems without changes to an app -- rather unlikely.