Add support for logins/sessions
This is a rather involved addition to make Keycard much more
self-sufficient as an authentication system. There are all the elements
here to manage multiple authentication strategies in a chain of command
and provide the usual conveniences for use in controllers.

There are three new primary terms:

 - Notary - the main component that manages identity verification
 - Certificate - the result of identity verification
 - Verification - an individual method for verifying identity

A Notary should be configured and made available to the application,
taking collaborators for the deployment mode (a request attribute
factory) and the verification methods.

The one complication is that a verification method needs to be able to
find and verify accounts. Rather than do this with model-based
conventions, each Verification can define its own contract for a method
it needs to call. An injected finder looks up the account, and in some
cases confirms parameters for the particular verification method. There
are some conveniences for binding to typical class methods, but there is
no requirement on the names or locations, only the signature.

Beyond that bit of configuration, the usage is rather similar to Sorcery
or Devise. There are the usual logged_in? and current_user methods and
an authenticate! method intended to be used as a before_action with a
rescue_from for missing/failed authentication handling. These controller
methods are available on a mixin module for familiarity.

Re: PFDR-208