Store timestamps in session as integers
Before, this was storing the timestamp as a Time object in the session.
However, the default Rails cookie session store can't round-trip the
Time object and instead returns it as a string, which causes the session
to get reset every time.

This instead stores the time as an integral number of seconds since
1970, which I have verified can be round-tripped through the Rails
cookie session store (testing that is out of scope here)